pyproject.toml
[build-system]
requires = ["setuptools>=77", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bcap"
readme = "README.md"
authors = []
license = {text = "GNU AGPL3"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
]
requires-python = ">=3.11"
dependencies = [
    "arches @ git+https://github.com/bcgov/arches@dev/8.1.x_bcgov",
    "arches-component-lab==0.0.1a9",
    "bcgov_arches_common @ git+https://github.com/bcgov/bcgov-arches-common@dev/2.0.x",
    "arches-querysets",
    "arches-controlled-lists @ git+https://github.com/archesproject/arches-controlled-lists@brf/feat/176_support_freeform_search",
    "boto3==1.26",
    "django-migrate-sql-deux==1.1.1",
    "django-pgtrigger==4.15.2",
    "django-storages==1.13",
    "redis",
    "python-dotenv",
    "oracledb",
    "html2text",
    "Authlib",
    "django-captcha"
]

version = "0.2.0"

[dependency-groups]
dev = [
    "livereload",
    "sst",
    "coverage",
    "django-silk==5.1.0",
    "pre-commit==3.8.0",
    "black==24.4.2",
    "requests-mock==1.12.1",
    "django-vite",
    "playwright",
    "pytest-playwright",
]

[tool.setuptools]
packages = ["bcap"]
